==> R/comb/03_subclusters.Rmd <==
---
title: "03 Subclusters - Combined Organelles"
author:
  - name: "Emir Turkes [emir.turkes@eturkes.com]"
  - name: "Diana Acosta [Diana.Acosta@osumc.edu]"
  - name: "Hongjun Fu [Hongjun.Fu@osumc.edu]"
  - name: "Ohio State University Wexner Medical Center"
date: '`r strftime(Sys.time(), "%B %d, %Y")`'
link-citations: true
output:
  html_document:
    code_folding: hide
    number_sections: true
    theme: lumen
    highlight: haddock
    toc: true
    toc_depth: 3
    toc_float:
      collapsed: false
      smooth_scroll: false
knit: (function(inputFile, encoding) {
  rmarkdown::render(inputFile, encoding = encoding, output_file = file.path(
    "..", "..", "results", unlist(strsplit(getwd(), "/"))[8], "03_subclusters.html"
  ))})
---

<style type="text/css">
body {font-size: 16px;}
h1.title {font-size: 35px;}
h1 {font-size: 24px;}
h2 {font-size: 22px;}
h3 {font-size: 20px;}
.toc-content {padding-left: 0px; padding-right: 0px;}
div.tocify {width: 100%;}
.tocify-subheader .tocify-item {font-size: 0.95em; padding-left: 25px; text-indent: 0;}
.tocify-subheader .tocify-subheader .tocify-item {
  font-size: 0.95em; padding-left: 35px; text-indent: 0;
}
div.main-container {max-width: none; width: 100%;}
</style>

*This file is a part of the [Acosta Organoid scRNAseq Project](https://github.com/eturkes/acosta-organoid-scRNAseq).*

In this document we subcluster annotated cell-types.
The data here will be referenced using the name ``r unlist(strsplit(getwd(), "/"))[8]``.

```{r}
#    This file is part of acosta-organoid-scRNAseq.
#    Copyright (C) 2022-2025  Emir Turkes, Diana Acosta, Gi Guo, Yuzhou Chang,
#    Hongjun Fu, Ohio State University Wexner Medical Center
#
#    This program is free software: you can redistribute it and/or modify
#    it under the terms of the GNU General Public License as published by
#    the Free Software Foundation, either version 3 of the License, or
#    (at your option) any later version.
#
#    This program is distributed in the hope that it will be useful,
#    but WITHOUT ANY WARRANTY; without even the implied warranty of
#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
#    GNU General Public License for more details.
#
#    You should have received a copy of the GNU General Public License
#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
#
#    Emir Turkes can be contacted at emir.turkes@eturkes.com

packages <- c(
  "conflicted", "Seurat", "ggplot2", "dplyr", "ggrepel", "SingleCellExperiment", "scuttle", "RCAv2",
  "data.table", "GSVA", "edgeR", "ComplexHeatmap", "ClusterMap", "filesstrings", "patchwork", "DT",
  "viridis", "pals"
)
invisible(suppressPackageStartupMessages(lapply(packages, library, character.only = TRUE)))
options(future.globals.maxSize = 10485760000)
source(file.path("..", "utils.R"))

protocol <- c("human", "droplet", "single-cell", "umis") # See `cluster_pipeline` in `utils.R`.
vars_to_regress <- NULL # See `cluster_pipeline`.
parallel_override <- 0 # See `parallel_plan` in `utils.R`.

# Metadata to plot after dimensionality reduction and clustering.
# Values in list can include "no_legend and/or "no_label" to exclude those.
# -------------------------------------------------------------------------
metadata_to_plot <- vector("list", 4)
names(metadata_to_plot) <- c("seurat_clusters", "sample", "timepoint", "genotype")
# -------------------------------------------------------------------------

analysis_no <- 3
data_name <- unlist(strsplit(getwd(), "/"))[8] # Name of dataset.
data_dir <- file.path("..", "..", "data") # Backed up data.
results_dir <- file.path("..", "..", "results")

# Unique cache directory for each analysis number.
# ------------------------------------------------
cache_dir <- file.path("..", "..", "cache", data_name, paste0("0", analysis_no))
if (!dir.exists(cache_dir)) {
  dir.create(cache_dir, recursive = TRUE)
}
# ------------------------------------------------

knitr::opts_chunk$set(fig.width = 10, fig.height = 7, dpi = 150)
```

# Prep

```{r}
seurat <- readRDS(file.path(cache_dir, "..", "02", "rca_seurat.rds"))
seurat

red_dim_plot(seurat, "rca_umap1", "rca_umap2", "cluster_names", "cat")

markers <- read.delim(
  file.path(data_dir, "gene-lists", "Human cortical development-gene-list summary - page3.tsv"),
  check.names = FALSE
)

markers <- markers[ , 1:which(colnames(markers) == "Polioudakis Neuron 2019")]
markers <- markers[ , -(which(colnames(markers) == "oRG vRG shared genes"))]
markers <- Filter(function(x) !all(is.na(x)), markers)
markers <- as.list(markers)
markers <- lapply(markers, function(x) x[!(x == "")])

remove <- c(
  "SORCS1", "ADRA2A", "SLA", "STMN2", "SORBS2", "ELAVL2", "DCX", "ZFHX4", "GLI3", "CD99", "IGFBP7",
  "AURKC", "HMGB1", "HMGB3", "CTNNB1", "HES1", "VIM", "PTN", "CLU", "IL33", "NPR3", "CCNB2",
  "KIF15", "ARHGAP11B", "CENPK", "NEUROD2", "NEUROD6", "DLX2", "ERBB4", "SOX5", "BCL11B"
)
markers <- markers
for (i in seq_along(remove)) {
  markers <- rapply(
    markers, function(x) ifelse(x == remove[i], "", x), how = "replace")
}
markers <- lapply(markers, function(x) x[!(x == "")])
remove <- which(
  names(markers) %in% c(
    "Radial Glia_vRG", "Radial Glia_tRG", "Microglia", "oRG specific genes", "vRG specific genes"
  )
)
markers <- markers[-remove]

final_annotations <- c()
final_annotations_long <- c()
cell_order <- c()
```

# Excitatory Neuron

## Standard Pipeline

```{r}
sub_seurat <- seurat[ , which(seurat$cluster_names == "Excitatory Neuron")]

DefaultAssay(sub_seurat) <- "RNA"
sub_seurat[["SCT"]] <- NULL
sub_name <- "Ex"
sub_seurat <- cluster_pipeline(
  sub_seurat, cache_dir, sub_name, protocol, vars_to_regress, parallel_override, FALSE, 1500
)
sub_seurat

for (i in 1:length(metadata_to_plot)) {
  print(names(metadata_to_plot)[i])
  if ("no_legend" %in% metadata_to_plot[[i]] && "no_label" %in% metadata_to_plot[[i]]) {
    print(red_dim_plot(sub_seurat, "umap1", "umap2", names(metadata_to_plot)[i]) + NoLegend())
  } else if ("no_legend" %in% metadata_to_plot[[i]]) {
    print(
      red_dim_plot(sub_seurat, "umap1", "umap2", names(metadata_to_plot)[i], "cat") + NoLegend()
    )
  } else if ("no_label" %in% metadata_to_plot[[i]]) {
    print(red_dim_plot(sub_seurat, "umap1", "umap2", names(metadata_to_plot)[i]))
  } else {
    print(red_dim_plot(sub_seurat, "umap1", "umap2", names(metadata_to_plot)[i], "cat"))
  }
}

red_dim_plot(sub_seurat, "umap1", "umap2", "subsets_mito_percent", "cont") +
  ggtitle("Mito Percentage")
red_dim_plot(sub_seurat, "umap1", "umap2", "subsets_ribo_percent", "cont") +
  ggtitle("Ribo Percentage")
```

## Separated By Timepoint

```{r}
timepoint_seurat <- SplitObject(sub_seurat, "timepoint")

for (i in seq_along(timepoint_seurat)) {
  DefaultAssay(timepoint_seurat[[i]]) <- "RNA"
  timepoint_seurat[[i]][["SCT"]] <- NULL
  save_name <- paste(sub_name, unique(timepoint_seurat[[i]]$timepoint), sep = "_")
  timepoint_seurat[[i]] <- cluster_pipeline(
    timepoint_seurat[[i]], cache_dir, save_name, protocol,
    vars_to_regress, parallel_override, FALSE, 1500
  )
  print(timepoint_seurat[[i]])

  for (j in 1:(length(metadata_to_plot))) {
    print(names(metadata_to_plot)[j])
    if ("no_legend" %in% metadata_to_plot[[j]] && "no_label" %in% metadata_to_plot[[j]]) {
      print(
        red_dim_plot(timepoint_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j]) +
          NoLegend()
      )
    } else if ("no_legend" %in% metadata_to_plot[[j]]) {
      print(
        red_dim_plot(timepoint_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j], "cat") +
          NoLegend()
      )
    } else if ("no_label" %in% metadata_to_plot[[j]]) {
      print(red_dim_plot(timepoint_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j]))
    } else {
      print(
        red_dim_plot(timepoint_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j], "cat")
      )
    }
  }
}
```

## Separated By Genotype

### Day 120

```{r}
D120_genotype_seurat <- subset(sub_seurat, timepoint == "D120")
D120_genotype_seurat <- SplitObject(D120_genotype_seurat, "genotype")

for (i in seq_along(D120_genotype_seurat)) {
  DefaultAssay(D120_genotype_seurat[[i]]) <- "RNA"
  D120_genotype_seurat[[i]][["SCT"]] <- NULL
  save_name <- paste(
    sub_name, unique(D120_genotype_seurat[[i]]$timepoint),
    unique(D120_genotype_seurat[[i]]$genotype), sep = "_"
  )
  D120_genotype_seurat[[i]] <- cluster_pipeline(
    D120_genotype_seurat[[i]], cache_dir, save_name, protocol,
    vars_to_regress, parallel_override, FALSE, 1500
  )
  print(D120_genotype_seurat[[i]])

  csv <- file.path(
    cache_dir,
    paste(
      sub_name, unique(D120_genotype_seurat[[i]]$timepoint),
      unique(D120_genotype_seurat[[i]]$genotype), "cluster_markers.csv",
      sep = "_"
    )
  )
  if (!file.exists(csv)) {
    cluster_markers <- FindAllMarkers(D120_genotype_seurat[[i]], only.pos = TRUE, verbose = FALSE)
    write.csv(cluster_markers, csv)
  }

  for (j in 1:(length(metadata_to_plot))) {
    print(names(metadata_to_plot)[j])
    if ("no_legend" %in% metadata_to_plot[[j]] && "no_label" %in% metadata_to_plot[[j]]) {
      print(
        red_dim_plot(D120_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j]) +
          NoLegend()
      )
    } else if ("no_legend" %in% metadata_to_plot[[j]]) {
      print(
        red_dim_plot(
          D120_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j], "cat"
        ) +
          NoLegend()
      )
    } else if ("no_label" %in% metadata_to_plot[[j]]) {
      print(red_dim_plot(D120_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j]))
    } else {
      print(
        red_dim_plot(D120_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j], "cat")
      )
    }
  }
}
```

### Day 72

```{r}
D72_genotype_seurat <- subset(sub_seurat, timepoint == "D72")
D72_genotype_seurat <- SplitObject(D72_genotype_seurat, "genotype")

for (i in seq_along(D72_genotype_seurat)) {
  DefaultAssay(D72_genotype_seurat[[i]]) <- "RNA"
  D72_genotype_seurat[[i]][["SCT"]] <- NULL
  save_name <- paste(
    sub_name, unique(D72_genotype_seurat[[i]]$timepoint),
    unique(D72_genotype_seurat[[i]]$genotype), sep = "_"
  )
  D72_genotype_seurat[[i]] <- cluster_pipeline(
    D72_genotype_seurat[[i]], cache_dir, save_name, protocol,
    vars_to_regress, parallel_override, FALSE, 1500
  )
  print(D72_genotype_seurat[[i]])

  csv <- file.path(
    cache_dir,
    paste(
      sub_name, unique(D72_genotype_seurat[[i]]$timepoint),
      unique(D72_genotype_seurat[[i]]$genotype), "cluster_markers.csv",
      sep = "_"
    )
  )
  if (file.exists(csv)) {
    cluster_markers <- read.csv(csv)
  } else {
    cluster_markers <- FindAllMarkers(D72_genotype_seurat[[i]], only.pos = TRUE, verbose = FALSE)
    write.csv(cluster_markers, csv)
  }

  for (j in 1:(length(metadata_to_plot))) {
    print(names(metadata_to_plot)[j])
    if ("no_legend" %in% metadata_to_plot[[j]] && "no_label" %in% metadata_to_plot[[j]]) {
      print(
        red_dim_plot(D72_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j]) +
          NoLegend()
      )
    } else if ("no_legend" %in% metadata_to_plot[[j]]) {
      print(
        red_dim_plot(
          D72_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j], "cat"
        ) +
          NoLegend()
      )
    } else if ("no_label" %in% metadata_to_plot[[j]]) {
      print(red_dim_plot(D72_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j]))
    } else {
      print(
        red_dim_plot(D72_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j], "cat")
      )
    }
  }
}
```

## ClusterMap

### Day 120

```{r}
marker_file_list <- c(
  HET = file.path(cache_dir, paste(sub_name, "D120", "HET", "cluster_markers.csv", sep = "_")),
  HOM = file.path(cache_dir, paste(sub_name, "D120", "HOM", "cluster_markers.csv", sep = "_")),
  WT = file.path(cache_dir, paste(sub_name, "D120", "WT", "cluster_markers.csv", sep = "_"))
)

file_list <- list(
  HET = D120_genotype_seurat[[1]], HOM = D120_genotype_seurat[[2]],
  WT = D120_genotype_seurat[[3]], comb = timepoint_seurat$D120
)
for (i in seq_along(file_list)) {
  file_list[[i]][["tsne"]] <- CreateDimReducObject(
    Embeddings(file_list[[i]], "umap"), key = "tsne_", assay = DefaultAssay(file_list[[i]])
  )
  if (names(file_list)[i] == "comb") {
    file_list[[i]] <- RenameCells(
      file_list[[i]], new.names = c(
        paste(names(file_list)[1], colnames(file_list[[1]]), sep = "_"),
        paste(names(file_list)[2], colnames(file_list[[2]]), sep = "_"),
        paste(names(file_list)[3], colnames(file_list[[3]]), sep = "_")
      )
    )
  }
}

out_dir <- file.path(
  results_dir, data_name, paste0("0", analysis_no, "-", sub_name, "-", "D120", "-", "clustermap")
)
if (!dir.exists(out_dir)) {
  dir.create(out_dir, recursive = TRUE)
  clustermap <- cluster_map(
    marker_file_list,
    output = file.path(out_dir, paste(sub_name, "D120", sep = ".")),
    single_obj_list = c(HET = file_list$HET, HOM = file_list$HOM, WT = file_list$WT),
    comb_obj = file_list$comb,
    comb_delim = "_",
    reduction = "umap",
    edge_cutoff = 0.01
  )
  files <- c(list.files(pattern = "*.png"), list.files(pattern = "*.pdf"))
  move_files(files, out_dir, TRUE)
}

clustermap_groups <- readRDS(
  file.path(out_dir, paste(sub_name, "D120", "new.group.list.RDS", sep = "."))
)
for (i in seq_along(clustermap_groups)) {
  file_list[[names(clustermap_groups)[i]]]$clustermap <- clustermap_groups[[i]]
  print(
    red_dim_plot(
      file_list[[names(clustermap_groups)[i]]], "umap1", "umap2", "clustermap", "cat"
    ) + ggtitle(names(clustermap_groups)[i])
  )
}
timepoint_seurat$D120$clustermap <- file_list$comb$clustermap
timepoint_seurat$D120@active.ident <- timepoint_seurat$D120$clustermap
file_list$comb@active.ident <- file_list$comb$clustermap

csv <- file.path(cache_dir, paste(sub_name, "D120", "clustermap", "markers.csv", sep = "_"))
if (file.exists(csv)) {
  clustermap_markers <- read.csv(csv, row.names = 1)
} else {
  clustermap_markers <- FindAllMarkers(file_list$comb, only.pos = TRUE, verbose = FALSE)
  write.csv(clustermap_markers, csv)
}

datatable_download(clustermap_markers)
```

### Day 72

```{r}
marker_file_list <- c(
  HET = file.path(cache_dir, paste(sub_name, "D72", "HET", "cluster_markers.csv", sep = "_")),
  HOM = file.path(cache_dir, paste(sub_name, "D72", "HOM", "cluster_markers.csv", sep = "_")),
  WT = file.path(cache_dir, paste(sub_name, "D72", "WT", "cluster_markers.csv", sep = "_"))
)

file_list <- list(
  HET = D72_genotype_seurat[[1]], HOM = D72_genotype_seurat[[2]],
  WT = D72_genotype_seurat[[3]], comb = timepoint_seurat$D72
)
for (i in seq_along(file_list)) {
  file_list[[i]][["tsne"]] <- CreateDimReducObject(
    Embeddings(file_list[[i]], "umap"), key = "tsne_", assay = DefaultAssay(file_list[[i]])
  )
  if (names(file_list)[i] == "comb") {
    file_list[[i]] <- RenameCells(
      file_list[[i]], new.names = c(
        paste(names(file_list)[1], colnames(file_list[[1]]), sep = "_"),
        paste(names(file_list)[2], colnames(file_list[[2]]), sep = "_"),
        paste(names(file_list)[3], colnames(file_list[[3]]), sep = "_")
      )
    )
  }
}

out_dir <- file.path(
  results_dir, data_name, paste0("0", analysis_no, "-", sub_name, "-", "D72", "-", "clustermap")
)
if (!dir.exists(out_dir)) {
  dir.create(out_dir, recursive = TRUE)
  clustermap <- cluster_map(
    marker_file_list,
    output = file.path(out_dir, paste(sub_name, "D72", sep = ".")),
    single_obj_list = c(HET = file_list$HET, HOM = file_list$HOM, WT = file_list$WT),
    comb_obj = file_list$comb,
    comb_delim = "_",
    reduction = "umap",
    edge_cutoff = 0.01
  )
  files <- c(list.files(pattern = "*.png"), list.files(pattern = "*.pdf"))
  move_files(files, out_dir, TRUE)
}

clustermap_groups <- readRDS(
  file.path(out_dir, paste(sub_name, "D72", "new.group.list.RDS", sep = "."))
)
for (i in seq_along(clustermap_groups)) {
  file_list[[names(clustermap_groups)[i]]]$clustermap <- clustermap_groups[[i]]
  print(
    red_dim_plot(
      file_list[[names(clustermap_groups)[i]]], "umap1", "umap2", "clustermap", "cat"
    ) + ggtitle(names(clustermap_groups)[i])
  )
}
timepoint_seurat$D72$clustermap <- file_list$comb$clustermap
timepoint_seurat$D72@active.ident <- timepoint_seurat$D72$clustermap
file_list$comb@active.ident <- file_list$comb$clustermap

csv <- file.path(cache_dir, paste(sub_name, "D72", "clustermap", "markers.csv", sep = "_"))
if (file.exists(csv)) {
  clustermap_markers <- read.csv(csv, row.names = 1)
} else {
  clustermap_markers <- FindAllMarkers(file_list$comb, only.pos = TRUE, verbose = FALSE)
  write.csv(clustermap_markers, csv)
}

datatable_download(clustermap_markers)
```

### All

```{r}
marker_file_list <- c(
  D120 = file.path(cache_dir, paste(sub_name, "D120", "clustermap_markers.csv", sep = "_")),
  D72 = file.path(cache_dir, paste(sub_name, "D72", "clustermap_markers.csv", sep = "_"))
)

file_list <- list(D120 = timepoint_seurat[[1]], D72 = timepoint_seurat[[2]], comb = sub_seurat)
for (i in seq_along(file_list)) {
  file_list[[i]][["tsne"]] <- CreateDimReducObject(
    Embeddings(file_list[[i]], "umap"), key = "tsne_", assay = DefaultAssay(file_list[[i]])
  )
  if (names(file_list)[i] == "comb") {
    file_list[[i]] <- RenameCells(
      file_list[[i]], new.names = c(
        paste(names(file_list)[1], colnames(file_list[[1]]), sep = "_"),
        paste(names(file_list)[2], colnames(file_list[[2]]), sep = "_")
      )
    )
  }
}

out_dir <- file.path(
  results_dir, data_name, paste0("0", analysis_no, "-", sub_name, "-", "clustermap")
)
if (!dir.exists(out_dir)) {
  dir.create(out_dir, recursive = TRUE)
  clustermap <- cluster_map(
    marker_file_list, output = file.path(out_dir, sub_name),
    single_obj_list = c(D120 = file_list$D120, D72 = file_list$D72), comb_obj = file_list$comb,
    comb_delim = "_", reduction = "umap",
    edge_cutoff = 0.01
  )
  files <- c(list.files(pattern = "*.png"), list.files(pattern = "*.pdf"))
  move_files(files, out_dir, TRUE)
}

clustermap_groups <- readRDS(file.path(out_dir, paste(sub_name, "new.group.list.RDS", sep = ".")))
for (i in seq_along(clustermap_groups)) {

  file_list[[names(clustermap_groups)[i]]]$clustermap <- clustermap_groups[[i]]

  print(
    red_dim_plot(
      file_list[[names(clustermap_groups)[i]]], "umap1", "umap2", "clustermap", "cat"
    ) + ggtitle(names(clustermap_groups)[i])
  )

  for (j in seq_along(markers)) {
    if (any(markers[[j]] %in% rownames(file_list[[names(clustermap_groups)[i]]]))) {
      print(
        FeaturePlot(
          file_list[[names(clustermap_groups)[i]]], markers[[j]],
          order = TRUE, cols = c("lightgrey", "red")
        ) +
          plot_annotation(
            names(markers)[j],
            theme = theme(plot.title = element_text(hjust = 0.5, size = 18, face = "bold"))
          )
      )
    }
  }

}
sub_seurat$clustermap <- file_list$comb$clustermap
sub_seurat@active.ident <- sub_seurat$clustermap

rds <- file.path(cache_dir, paste(sub_name, "clustermap_markers.rds", sep = "_"))
if (file.exists(rds)) {
  clustermap_markers <- readRDS(rds)
} else {
  clustermap_markers <- FindAllMarkers(sub_seurat, only.pos = TRUE, verbose = FALSE)
  saveRDS(clustermap_markers, rds, compress = FALSE)
}

datatable_download(clustermap_markers)
```

## Annotations

```{r}
file_list$D120@active.ident <- file_list$D120$clustermap
file_list$D120 <- RenameIdents(
  file_list$D120,
  "4" = "Ex Newborn 01", "16" = "Ex Newborn 02", "2" = "Ex Upper Layer 01",
  "1" = "Ex Upper Layer 02", "11" = "Ex Upper Layer 03", "13" = "Ex Upper Layer 04",
  "5" = "Ex Deep Layer 01", "18" = "Ex Deep Layer 02", "17" = "Ex Deep Layer 03",
  "3" = "Ex Deep Layer 04", "6" = "Ex Deep Layer 05", "7" = "Ex Upper Layer 06",
  "12" = "Ex Deep Layer 06", "9" = "Ex Upper Layer 07", "15" = "Ex Upper Layer 05",
  "8" = "Ex Newborn 03", "14" = "Ex Unclassified 01", "10" = "Ex Unclassified 02"
)
file_list$D120$clustermap_names <- factor(as.character(file_list$D120@active.ident))
file_list$D120$clustermap_names_short <- factor(
  sub("\\s+[^ ]+$", "", as.character(file_list$D120$clustermap_names))
)
red_dim_plot(file_list$D120, "umap1", "umap2", "clustermap_names", "cat") + ggtitle("D120")
red_dim_plot(file_list$D120, "umap1", "umap2", "clustermap_names_short", "cat") + ggtitle("D120")

file_list$D72@active.ident <- file_list$D72$clustermap
file_list$D72 <- RenameIdents(
  file_list$D72,
  "4" = "Ex Newborn 01", "16" = "Ex Newborn 02", "2" = "Ex Upper Layer 01",
  "1" = "Ex Upper Layer 02", "11" = "Ex Upper Layer 03", "13" = "Ex Upper Layer 04",
  "5" = "Ex Deep Layer 01", "18" = "Ex Deep Layer 02", "17" = "Ex Deep Layer 03",
  "3" = "Ex Deep Layer 04", "6" = "Ex Deep Layer 05", "7" = "Ex Upper Layer 06",
  "12" = "Ex Deep Layer 06", "9" = "Ex Upper Layer 07", "15" = "Ex Upper Layer 05",
  "8" = "Ex Newborn 03", "14" = "Ex Unclassified 01", "10" = "Ex Unclassified 02"
)
file_list$D72$clustermap_names <- factor(as.character(file_list$D72@active.ident))
file_list$D72$clustermap_names_short <- factor(
  sub("\\s+[^ ]+$", "", as.character(file_list$D72$clustermap_names))
)
red_dim_plot(file_list$D72, "umap1", "umap2", "clustermap_names", "cat") + ggtitle("D72")
red_dim_plot(file_list$D72, "umap1", "umap2", "clustermap_names_short", "cat") + ggtitle("D72")

file_list$comb@active.ident <- file_list$comb$clustermap
file_list$comb <- RenameIdents(
  file_list$comb,
  "4" = "Ex Newborn 01", "16" = "Ex Newborn 02", "2" = "Ex Upper Layer 01",
  "1" = "Ex Upper Layer 02", "11" = "Ex Upper Layer 03", "13" = "Ex Upper Layer 04",
  "5" = "Ex Deep Layer 01", "18" = "Ex Deep Layer 02", "17" = "Ex Deep Layer 03",
  "3" = "Ex Deep Layer 04", "6" = "Ex Deep Layer 05", "7" = "Ex Upper Layer 06",
  "12" = "Ex Deep Layer 06", "9" = "Ex Upper Layer 07", "15" = "Ex Upper Layer 05",
  "8" = "Ex Newborn 03", "14" = "Ex Unclassified 01", "10" = "Ex Unclassified 02"
)
file_list$comb$clustermap_names <- factor(as.character(file_list$comb@active.ident))
file_list$comb$clustermap_names_short <- factor(
  sub("\\s+[^ ]+$", "", as.character(file_list$comb$clustermap_names))
)
red_dim_plot(file_list$comb, "umap1", "umap2", "clustermap_names", "cat") + ggtitle("D120 + D72")
red_dim_plot(file_list$comb, "umap1", "umap2", "clustermap_names_short", "cat") +
  ggtitle("D120 + D72")

rds <- file.path(cache_dir, paste(sub_name, "seurat_list.rds", sep = "_"))
if (!file.exists(rds)) {
  saveRDS(file_list, rds, compress = FALSE)
}

final_annotations <- c(final_annotations, as.character(file_list$comb$clustermap_names_short))
final_annotations_long <- c(final_annotations_long, as.character(file_list$comb$clustermap_names))
cell_order <- c(cell_order, colnames(sub_seurat))
```

# Radial Glia

## Standard Pipeline

```{r}
sub_seurat <- seurat[ , which(seurat$cluster_names == "Radial Glia")]

DefaultAssay(sub_seurat) <- "RNA"
sub_seurat[["SCT"]] <- NULL
sub_name <- "RG"
sub_seurat <- cluster_pipeline(
  sub_seurat, cache_dir, sub_name, protocol, vars_to_regress, parallel_override, FALSE, 500
)
sub_seurat

for (i in 1:length(metadata_to_plot)) {
  print(names(metadata_to_plot)[i])
  if ("no_legend" %in% metadata_to_plot[[i]] && "no_label" %in% metadata_to_plot[[i]]) {
    print(red_dim_plot(sub_seurat, "umap1", "umap2", names(metadata_to_plot)[i]) + NoLegend())
  } else if ("no_legend" %in% metadata_to_plot[[i]]) {
    print(
      red_dim_plot(sub_seurat, "umap1", "umap2", names(metadata_to_plot)[i], "cat") + NoLegend()
    )
  } else if ("no_label" %in% metadata_to_plot[[i]]) {
    print(red_dim_plot(sub_seurat, "umap1", "umap2", names(metadata_to_plot)[i]))
  } else {
    print(red_dim_plot(sub_seurat, "umap1", "umap2", names(metadata_to_plot)[i], "cat"))
  }
}

red_dim_plot(sub_seurat, "umap1", "umap2", "subsets_mito_percent", "cont") +
  ggtitle("Mito Percentage")
red_dim_plot(sub_seurat, "umap1", "umap2", "subsets_ribo_percent", "cont") +
  ggtitle("Ribo Percentage")
```

## Separated By Timepoint

```{r}
timepoint_seurat <- SplitObject(sub_seurat, "timepoint")

for (i in seq_along(timepoint_seurat)) {
  DefaultAssay(timepoint_seurat[[i]]) <- "RNA"
  timepoint_seurat[[i]][["SCT"]] <- NULL
  save_name <- paste(sub_name, unique(timepoint_seurat[[i]]$timepoint), sep = "_")
  timepoint_seurat[[i]] <- cluster_pipeline(
    timepoint_seurat[[i]], cache_dir, save_name, protocol,
    vars_to_regress, parallel_override, FALSE, 500
  )
  print(timepoint_seurat[[i]])

  for (j in 1:(length(metadata_to_plot))) {
    print(names(metadata_to_plot)[j])
    if ("no_legend" %in% metadata_to_plot[[j]] && "no_label" %in% metadata_to_plot[[j]]) {
      print(
        red_dim_plot(timepoint_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j]) +
          NoLegend()
      )
    } else if ("no_legend" %in% metadata_to_plot[[j]]) {
      print(
        red_dim_plot(timepoint_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j], "cat") +
          NoLegend()
      )
    } else if ("no_label" %in% metadata_to_plot[[j]]) {
      print(red_dim_plot(timepoint_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j]))
    } else {
      print(
        red_dim_plot(timepoint_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j], "cat")
      )
    }
  }
}
```

## Separated By Genotype

### Day 120

```{r}
D120_genotype_seurat <- subset(sub_seurat, timepoint == "D120")
D120_genotype_seurat <- SplitObject(D120_genotype_seurat, "genotype")

for (i in seq_along(D120_genotype_seurat)) {
  DefaultAssay(D120_genotype_seurat[[i]]) <- "RNA"
  D120_genotype_seurat[[i]][["SCT"]] <- NULL
  save_name <- paste(
    sub_name, unique(D120_genotype_seurat[[i]]$timepoint),
    unique(D120_genotype_seurat[[i]]$genotype), sep = "_"
  )
  D120_genotype_seurat[[i]] <- cluster_pipeline(
    D120_genotype_seurat[[i]], cache_dir, save_name, protocol,
    vars_to_regress, parallel_override, FALSE, 500
  )
  print(D120_genotype_seurat[[i]])

  csv <- file.path(
    cache_dir,
    paste(
      sub_name, unique(D120_genotype_seurat[[i]]$timepoint),
      unique(D120_genotype_seurat[[i]]$genotype), "cluster_markers.csv",
      sep = "_"
    )
  )
  if (!file.exists(csv)) {
    cluster_markers <- FindAllMarkers(D120_genotype_seurat[[i]], only.pos = TRUE, verbose = FALSE)
    write.csv(cluster_markers, csv)
  }

  for (j in 1:(length(metadata_to_plot))) {
    print(names(metadata_to_plot)[j])
    if ("no_legend" %in% metadata_to_plot[[j]] && "no_label" %in% metadata_to_plot[[j]]) {
      print(
        red_dim_plot(D120_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j]) +
          NoLegend()
      )
    } else if ("no_legend" %in% metadata_to_plot[[j]]) {
      print(
        red_dim_plot(
          D120_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j], "cat"
        ) +
          NoLegend()
      )
    } else if ("no_label" %in% metadata_to_plot[[j]]) {
      print(red_dim_plot(D120_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j]))
    } else {
      print(
        red_dim_plot(D120_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j], "cat")
      )
    }
  }
}
```

### Day 72

```{r}
D72_genotype_seurat <- subset(sub_seurat, timepoint == "D72")
D72_genotype_seurat <- SplitObject(D72_genotype_seurat, "genotype")

for (i in seq_along(D72_genotype_seurat)) {
  DefaultAssay(D72_genotype_seurat[[i]]) <- "RNA"
  D72_genotype_seurat[[i]][["SCT"]] <- NULL
  save_name <- paste(
    sub_name, unique(D72_genotype_seurat[[i]]$timepoint),
    unique(D72_genotype_seurat[[i]]$genotype), sep = "_"
  )
  D72_genotype_seurat[[i]] <- cluster_pipeline(
    D72_genotype_seurat[[i]], cache_dir, save_name, protocol,
    vars_to_regress, parallel_override, FALSE, 500
  )
  print(D72_genotype_seurat[[i]])

  csv <- file.path(
    cache_dir,
    paste(
      sub_name, unique(D72_genotype_seurat[[i]]$timepoint),
      unique(D72_genotype_seurat[[i]]$genotype), "cluster_markers.csv",
      sep = "_"
    )
  )
  if (file.exists(csv)) {
    cluster_markers <- read.csv(csv)
  } else {
    cluster_markers <- FindAllMarkers(D72_genotype_seurat[[i]], only.pos = TRUE, verbose = FALSE)
    write.csv(cluster_markers, csv)
  }

  for (j in 1:(length(metadata_to_plot))) {
    print(names(metadata_to_plot)[j])
    if ("no_legend" %in% metadata_to_plot[[j]] && "no_label" %in% metadata_to_plot[[j]]) {
      print(
        red_dim_plot(D72_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j]) +
          NoLegend()
      )
    } else if ("no_legend" %in% metadata_to_plot[[j]]) {
      print(
        red_dim_plot(
          D72_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j], "cat"
        ) +
          NoLegend()
      )
    } else if ("no_label" %in% metadata_to_plot[[j]]) {
      print(red_dim_plot(D72_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j]))
    } else {
      print(
        red_dim_plot(D72_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j], "cat")
      )
    }
  }
}
```

## ClusterMap

### Day 120

```{r}
marker_file_list <- c(
  HET = file.path(cache_dir, paste(sub_name, "D120", "HET", "cluster_markers.csv", sep = "_")),
  HOM = file.path(cache_dir, paste(sub_name, "D120", "HOM", "cluster_markers.csv", sep = "_")),
  WT = file.path(cache_dir, paste(sub_name, "D120", "WT", "cluster_markers.csv", sep = "_"))
)

file_list <- list(
  HET = D120_genotype_seurat[[1]], HOM = D120_genotype_seurat[[2]],
  WT = D120_genotype_seurat[[3]], comb = timepoint_seurat$D120
)
for (i in seq_along(file_list)) {
  file_list[[i]][["tsne"]] <- CreateDimReducObject(
    Embeddings(file_list[[i]], "umap"), key = "tsne_", assay = DefaultAssay(file_list[[i]])
  )
  if (names(file_list)[i] == "comb") {
    file_list[[i]] <- RenameCells(
      file_list[[i]], new.names = c(
        paste(names(file_list)[1], colnames(file_list[[1]]), sep = "_"),
        paste(names(file_list)[2], colnames(file_list[[2]]), sep = "_"),
        paste(names(file_list)[3], colnames(file_list[[3]]), sep = "_")
      )
    )
  }
}

out_dir <- file.path(
  results_dir, data_name, paste0("0", analysis_no, "-", sub_name, "-", "D120", "-", "clustermap")
)
if (!dir.exists(out_dir)) {
  dir.create(out_dir, recursive = TRUE)
  clustermap <- cluster_map(
    marker_file_list,
    output = file.path(out_dir, paste(sub_name, "D120", sep = ".")),
    single_obj_list = c(HET = file_list$HET, HOM = file_list$HOM, WT = file_list$WT),
    comb_obj = file_list$comb,
    comb_delim = "_",
    reduction = "umap",
    edge_cutoff = 0.01
  )
  files <- c(list.files(pattern = "*.png"), list.files(pattern = "*.pdf"))
  move_files(files, out_dir, TRUE)
}

clustermap_groups <- readRDS(
  file.path(out_dir, paste(sub_name, "D120", "new.group.list.RDS", sep = "."))
)
for (i in seq_along(clustermap_groups)) {
  file_list[[names(clustermap_groups)[i]]]$clustermap <- clustermap_groups[[i]]
  print(
    red_dim_plot(
      file_list[[names(clustermap_groups)[i]]], "umap1", "umap2", "clustermap", "cat"
    ) + ggtitle(names(clustermap_groups)[i])
  )
}
timepoint_seurat$D120$clustermap <- file_list$comb$clustermap
timepoint_seurat$D120@active.ident <- timepoint_seurat$D120$clustermap
file_list$comb@active.ident <- file_list$comb$clustermap

csv <- file.path(cache_dir, paste(sub_name, "D120", "clustermap", "markers.csv", sep = "_"))
if (file.exists(csv)) {
  clustermap_markers <- read.csv(csv, row.names = 1)
} else {
  clustermap_markers <- FindAllMarkers(file_list$comb, only.pos = TRUE, verbose = FALSE)
  write.csv(clustermap_markers, csv)
}

datatable_download(clustermap_markers)
```

### Day 72

```{r}
marker_file_list <- c(
  HET = file.path(cache_dir, paste(sub_name, "D72", "HET", "cluster_markers.csv", sep = "_")),
  HOM = file.path(cache_dir, paste(sub_name, "D72", "HOM", "cluster_markers.csv", sep = "_")),
  WT = file.path(cache_dir, paste(sub_name, "D72", "WT", "cluster_markers.csv", sep = "_"))
)

file_list <- list(
  HET = D72_genotype_seurat[[1]], HOM = D72_genotype_seurat[[2]],
  WT = D72_genotype_seurat[[3]], comb = timepoint_seurat$D72
)
for (i in seq_along(file_list)) {
  file_list[[i]][["tsne"]] <- CreateDimReducObject(
    Embeddings(file_list[[i]], "umap"), key = "tsne_", assay = DefaultAssay(file_list[[i]])
  )
  if (names(file_list)[i] == "comb") {
    file_list[[i]] <- RenameCells(
      file_list[[i]], new.names = c(
        paste(names(file_list)[1], colnames(file_list[[1]]), sep = "_"),
        paste(names(file_list)[2], colnames(file_list[[2]]), sep = "_"),
        paste(names(file_list)[3], colnames(file_list[[3]]), sep = "_")
      )
    )
  }
}

out_dir <- file.path(
  results_dir, data_name, paste0("0", analysis_no, "-", sub_name, "-", "D72", "-", "clustermap")
)
if (!dir.exists(out_dir)) {
  dir.create(out_dir, recursive = TRUE)
  clustermap <- cluster_map(
    marker_file_list,
    output = file.path(out_dir, paste(sub_name, "D72", sep = ".")),
    single_obj_list = c(HET = file_list$HET, HOM = file_list$HOM, WT = file_list$WT),
    comb_obj = file_list$comb,
    comb_delim = "_",
    reduction = "umap",
    edge_cutoff = 0.01
  )
  files <- c(list.files(pattern = "*.png"), list.files(pattern = "*.pdf"))
  move_files(files, out_dir, TRUE)
}

clustermap_groups <- readRDS(
  file.path(out_dir, paste(sub_name, "D72", "new.group.list.RDS", sep = "."))
)
for (i in seq_along(clustermap_groups)) {
  file_list[[names(clustermap_groups)[i]]]$clustermap <- clustermap_groups[[i]]
  print(
    red_dim_plot(
      file_list[[names(clustermap_groups)[i]]], "umap1", "umap2", "clustermap", "cat"
    ) + ggtitle(names(clustermap_groups)[i])
  )
}
timepoint_seurat$D72$clustermap <- file_list$comb$clustermap
timepoint_seurat$D72@active.ident <- timepoint_seurat$D72$clustermap
file_list$comb@active.ident <- file_list$comb$clustermap

csv <- file.path(cache_dir, paste(sub_name, "D72", "clustermap", "markers.csv", sep = "_"))
if (file.exists(csv)) {
  clustermap_markers <- read.csv(csv, row.names = 1)
} else {
  clustermap_markers <- FindAllMarkers(file_list$comb, only.pos = TRUE, verbose = FALSE)
  write.csv(clustermap_markers, csv)
}

datatable_download(clustermap_markers)
```

### All

```{r}
marker_file_list <- c(
  D120 = file.path(cache_dir, paste(sub_name, "D120", "clustermap_markers.csv", sep = "_")),
  D72 = file.path(cache_dir, paste(sub_name, "D72", "clustermap_markers.csv", sep = "_"))
)

file_list <- list(D120 = timepoint_seurat[[1]], D72 = timepoint_seurat[[2]], comb = sub_seurat)
for (i in seq_along(file_list)) {
  file_list[[i]][["tsne"]] <- CreateDimReducObject(
    Embeddings(file_list[[i]], "umap"), key = "tsne_", assay = DefaultAssay(file_list[[i]])
  )
  if (names(file_list)[i] == "comb") {
    file_list[[i]] <- RenameCells(
      file_list[[i]], new.names = c(
        paste(names(file_list)[1], colnames(file_list[[1]]), sep = "_"),
        paste(names(file_list)[2], colnames(file_list[[2]]), sep = "_")
      )
    )
  }
}

out_dir <- file.path(
  results_dir, data_name, paste0("0", analysis_no, "-", sub_name, "-", "clustermap")
)
if (!dir.exists(out_dir)) {
  dir.create(out_dir, recursive = TRUE)
  clustermap <- cluster_map(
    marker_file_list, output = file.path(out_dir, sub_name),
    single_obj_list = c(D120 = file_list$D120, D72 = file_list$D72), comb_obj = file_list$comb,
    comb_delim = "_", reduction = "umap",
    edge_cutoff = 0.01
  )
  files <- c(list.files(pattern = "*.png"), list.files(pattern = "*.pdf"))
  move_files(files, out_dir, TRUE)
}

clustermap_groups <- readRDS(file.path(out_dir, paste(sub_name, "new.group.list.RDS", sep = ".")))
for (i in seq_along(clustermap_groups)) {

  file_list[[names(clustermap_groups)[i]]]$clustermap <- clustermap_groups[[i]]

  print(
    red_dim_plot(
      file_list[[names(clustermap_groups)[i]]], "umap1", "umap2", "clustermap", "cat"
    ) + ggtitle(names(clustermap_groups)[i])
  )

  for (j in seq_along(markers)) {
    if (any(markers[[j]] %in% rownames(file_list[[names(clustermap_groups)[i]]]))) {
      print(
        FeaturePlot(
          file_list[[names(clustermap_groups)[i]]], markers[[j]],
          order = TRUE, cols = c("lightgrey", "red")
        ) +
          plot_annotation(
            names(markers)[j],
            theme = theme(plot.title = element_text(hjust = 0.5, size = 18, face = "bold"))
          )
      )
    }
  }

}
sub_seurat$clustermap <- file_list$comb$clustermap
sub_seurat@active.ident <- sub_seurat$clustermap

rds <- file.path(cache_dir, paste(sub_name, "clustermap_markers.rds", sep = "_"))
if (file.exists(rds)) {
  clustermap_markers <- readRDS(rds)
} else {
  clustermap_markers <- FindAllMarkers(sub_seurat, only.pos = TRUE, verbose = FALSE)
  saveRDS(clustermap_markers, rds, compress = FALSE)
}

datatable_download(clustermap_markers)
```

## Annotations

```{r}
file_list$D120@active.ident <- file_list$D120$clustermap
file_list$D120 <- RenameIdents(
  file_list$D120,
  "12" = "RG Astrocyte 02", "8" = "RG Astrocyte 01", "30" = "RG Astrocyte 03",
  "38" = "RG Astrocyte 04", "28" = "RG OPC 01", "15" = "RG OPC 03",
  "2" = "RG Early 01", "4" = "RG Early 02", "7" = "RG Early 03",
  "16" = "RG Early 04", "32" = "RG Early 05", "33" = "RG Early 06",
  "54" = "RG Early 07", "60" = "RG Early 08", "50" = "RG Early 09",
  "61" = "RG Early 10", "44" = "RG Early 11", "42" = "RG Early 14",
  "14" = "RG Late 01", "21" = "RG Late 02", "11" = "RG Late 03",
  "1" = "RG Late 04", "6" = "RG Late 05", "27" = "RG Late 06",
  "29" = "RG Astrocyte 05", "20" = "RG Astrocyte 06", "34" = "RG Late 07",
  "24" = "RG Unclassified 05", "25" = "RG Late 08", "26" = "RG Astrocyte 07",
  "36" = "RG Late 09", "17" = "RG Late 10", "59" = "RG Late 11",
  "58" = "RG Late 12", "53" = "RG Late 13", "18" = "RG Late 14",
  "51" = "RG Late 15", "62" = "RG Late 16", "56" = "RG Late 17",
  "45" = "RG Late 18", "55" = "RG Late 19", "41" = "RG Late 20",
  "37" = "RG Late 21", "19" = "RG Late 22", "39" = "RG Early 12",
  "9" = "RG Early 13", "13" = "RG Astrocyte 08", "31" = "RG OPC 02",
  "3" = "RG Unclassified 02", "10" = "RG Unclassified 03", "35" = "RG Unclassified 04",
  "43" = "RG Unclassified 06", "46" = "RG Unclassified 07", "49" = "RG Unclassified 08",
  "40" = "RG Unclassified 09", "47" = "RG Unclassified 10", "38" = "RG Unclassified 11",
  "48" = "RG Unclassified 12", "24" = "RG Unclassified 13", "57" = "RG Unclassified 14",
  "51" = "RG Unclassified 15", "22" = "RG Unclassified 16", "52" = "RG Unclassified 17",
  "23" = "RG Unclassified 18", "5" = "RG Unclassified 01"
)
file_list$D120$clustermap_names <- factor(as.character(file_list$D120@active.ident))
file_list$D120$clustermap_names_short <- factor(
  sub("\\s+[^ ]+$", "", as.character(file_list$D120$clustermap_names))
)
red_dim_plot(file_list$D120, "umap1", "umap2", "clustermap_names", "cat") + ggtitle("D120")
red_dim_plot(file_list$D120, "umap1", "umap2", "clustermap_names_short", "cat") + ggtitle("D120")

file_list$D72@active.ident <- file_list$D72$clustermap
file_list$D72 <- RenameIdents(
  file_list$D72,
  "12" = "RG Astrocyte 02", "8" = "RG Astrocyte 01", "30" = "RG Astrocyte 03",
  "38" = "RG Astrocyte 04", "28" = "RG OPC 01", "15" = "RG OPC 03",
  "2" = "RG Early 01", "4" = "RG Early 02", "7" = "RG Early 03",
  "16" = "RG Early 04", "32" = "RG Early 05", "33" = "RG Early 06",
  "54" = "RG Early 07", "60" = "RG Early 08", "50" = "RG Early 09",
  "61" = "RG Early 10", "44" = "RG Early 11", "42" = "RG Early 14",
  "14" = "RG Late 01", "21" = "RG Late 02", "11" = "RG Late 03",
  "1" = "RG Late 04", "6" = "RG Late 05", "27" = "RG Late 06",
  "29" = "RG Astrocyte 05", "20" = "RG Astrocyte 06", "34" = "RG Late 07",
  "24" = "RG Unclassified 05", "25" = "RG Late 08", "26" = "RG Astrocyte 07",
  "36" = "RG Late 09", "17" = "RG Late 10", "59" = "RG Late 11",
  "58" = "RG Late 12", "53" = "RG Late 13", "18" = "RG Late 14",
  "51" = "RG Late 15", "62" = "RG Late 16", "56" = "RG Late 17",
  "45" = "RG Late 18", "55" = "RG Late 19", "41" = "RG Late 20",
  "37" = "RG Late 21", "19" = "RG Late 22", "39" = "RG Early 12",
  "9" = "RG Early 13", "13" = "RG Astrocyte 08", "31" = "RG OPC 02",
  "3" = "RG Unclassified 02", "10" = "RG Unclassified 03", "35" = "RG Unclassified 04",
  "43" = "RG Unclassified 06", "46" = "RG Unclassified 07", "49" = "RG Unclassified 08",
  "40" = "RG Unclassified 09", "47" = "RG Unclassified 10", "38" = "RG Unclassified 11",
  "48" = "RG Unclassified 12", "24" = "RG Unclassified 13", "57" = "RG Unclassified 14",
  "51" = "RG Unclassified 15", "22" = "RG Unclassified 16", "52" = "RG Unclassified 17",
  "23" = "RG Unclassified 18", "5" = "RG Unclassified 01"
)
file_list$D72$clustermap_names <- factor(as.character(file_list$D72@active.ident))
file_list$D72$clustermap_names_short <- factor(
  sub("\\s+[^ ]+$", "", as.character(file_list$D72$clustermap_names))
)
red_dim_plot(file_list$D72, "umap1", "umap2", "clustermap_names", "cat") + ggtitle("D72")
red_dim_plot(file_list$D72, "umap1", "umap2", "clustermap_names_short", "cat") + ggtitle("D72")

file_list$comb@active.ident <- file_list$comb$clustermap
file_list$comb <- RenameIdents(
  file_list$comb,
  "12" = "RG Astrocyte 02", "8" = "RG Astrocyte 01", "30" = "RG Astrocyte 03",
  "38" = "RG Astrocyte 04", "28" = "RG OPC 01", "15" = "RG OPC 03",
  "2" = "RG Early 01", "4" = "RG Early 02", "7" = "RG Early 03",
  "16" = "RG Early 04", "32" = "RG Early 05", "33" = "RG Early 06",
  "54" = "RG Early 07", "60" = "RG Early 08", "50" = "RG Early 09",
  "61" = "RG Early 10", "44" = "RG Early 11", "42" = "RG Early 14",
  "14" = "RG Late 01", "21" = "RG Late 02", "11" = "RG Late 03",
  "1" = "RG Late 04", "6" = "RG Late 05", "27" = "RG Late 06",
  "29" = "RG Astrocyte 05", "20" = "RG Astrocyte 06", "34" = "RG Late 07",
  "24" = "RG Unclassified 05", "25" = "RG Late 08", "26" = "RG Astrocyte 07",
  "36" = "RG Late 09", "17" = "RG Late 10", "59" = "RG Late 11",
  "58" = "RG Late 12", "53" = "RG Late 13", "18" = "RG Late 14",
  "51" = "RG Late 15", "62" = "RG Late 16", "56" = "RG Late 17",
  "45" = "RG Late 18", "55" = "RG Late 19", "41" = "RG Late 20",
  "37" = "RG Late 21", "19" = "RG Late 22", "39" = "RG Early 12",
  "9" = "RG Early 13", "13" = "RG Astrocyte 08", "31" = "RG OPC 02",
  "3" = "RG Unclassified 02", "10" = "RG Unclassified 03", "35" = "RG Unclassified 04",
  "43" = "RG Unclassified 06", "46" = "RG Unclassified 07", "49" = "RG Unclassified 08",
  "40" = "RG Unclassified 09", "47" = "RG Unclassified 10", "38" = "RG Unclassified 11",
  "48" = "RG Unclassified 12", "24" = "RG Unclassified 13", "57" = "RG Unclassified 14",
  "51" = "RG Unclassified 15", "22" = "RG Unclassified 16", "52" = "RG Unclassified 17",
  "23" = "RG Unclassified 18", "5" = "RG Unclassified 01"
)
file_list$comb$clustermap_names <- factor(as.character(file_list$comb@active.ident))
file_list$comb$clustermap_names_short <- factor(
  sub("\\s+[^ ]+$", "", as.character(file_list$comb$clustermap_names))
)
red_dim_plot(file_list$comb, "umap1", "umap2", "clustermap_names", "cat") + ggtitle("D120 + D72")
red_dim_plot(file_list$comb, "umap1", "umap2", "clustermap_names_short", "cat") +
  ggtitle("D120 + D72")

rds <- file.path(cache_dir, paste(sub_name, "seurat_list.rds", sep = "_"))
if (!file.exists(rds)) {
  saveRDS(file_list, rds, compress = FALSE)
}

final_annotations <- c(final_annotations, as.character(file_list$comb$clustermap_names_short))
final_annotations_long <- c(final_annotations_long, as.character(file_list$comb$clustermap_names))
cell_order <- c(cell_order, colnames(sub_seurat))
```

# IPC

## Standard Pipeline

```{r}
sub_seurat <- seurat[ , which(seurat$cluster_names == "IPC")]

DefaultAssay(sub_seurat) <- "RNA"
sub_seurat[["SCT"]] <- NULL
sub_name <- "IPC"
sub_seurat <- cluster_pipeline(
  sub_seurat, cache_dir, sub_name, protocol, vars_to_regress, parallel_override, FALSE
)
sub_seurat

for (i in 1:length(metadata_to_plot)) {
  print(names(metadata_to_plot)[i])
  if ("no_legend" %in% metadata_to_plot[[i]] && "no_label" %in% metadata_to_plot[[i]]) {
    print(red_dim_plot(sub_seurat, "umap1", "umap2", names(metadata_to_plot)[i]) + NoLegend())
  } else if ("no_legend" %in% metadata_to_plot[[i]]) {
    print(
      red_dim_plot(sub_seurat, "umap1", "umap2", names(metadata_to_plot)[i], "cat") + NoLegend()
    )
  } else if ("no_label" %in% metadata_to_plot[[i]]) {
    print(red_dim_plot(sub_seurat, "umap1", "umap2", names(metadata_to_plot)[i]))
  } else {
    print(red_dim_plot(sub_seurat, "umap1", "umap2", names(metadata_to_plot)[i], "cat"))
  }
}

red_dim_plot(sub_seurat, "umap1", "umap2", "subsets_mito_percent", "cont") +
  ggtitle("Mito Percentage")
red_dim_plot(sub_seurat, "umap1", "umap2", "subsets_ribo_percent", "cont") +
  ggtitle("Ribo Percentage")
```

## Separated By Timepoint

```{r}
timepoint_seurat <- SplitObject(sub_seurat, "timepoint")

for (i in seq_along(timepoint_seurat)) {
  DefaultAssay(timepoint_seurat[[i]]) <- "RNA"
  timepoint_seurat[[i]][["SCT"]] <- NULL
  save_name <- paste(sub_name, unique(timepoint_seurat[[i]]$timepoint), sep = "_")
  timepoint_seurat[[i]] <- cluster_pipeline(
    timepoint_seurat[[i]], cache_dir, save_name, protocol, vars_to_regress, parallel_override, FALSE
  )
  print(timepoint_seurat[[i]])

  for (j in 1:(length(metadata_to_plot))) {
    print(names(metadata_to_plot)[j])
    if ("no_legend" %in% metadata_to_plot[[j]] && "no_label" %in% metadata_to_plot[[j]]) {
      print(
        red_dim_plot(timepoint_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j]) +
          NoLegend()
      )
    } else if ("no_legend" %in% metadata_to_plot[[j]]) {
      print(
        red_dim_plot(timepoint_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j], "cat") +
          NoLegend()
      )
    } else if ("no_label" %in% metadata_to_plot[[j]]) {
      print(red_dim_plot(timepoint_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j]))
    } else {
      print(
        red_dim_plot(timepoint_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j], "cat")
      )
    }
  }

  for (j in seq_along(markers)) {
    if (any(markers[[j]] %in% rownames(timepoint_seurat[[i]]))) {
      print(
        FeaturePlot(
          timepoint_seurat[[i]], markers[[j]],
          order = TRUE, cols = c("lightgrey", "red")
        ) +
          plot_annotation(
            names(markers)[j],
            theme = theme(plot.title = element_text(hjust = 0.5, size = 18, face = "bold"))
          )
      )
    }
  }
}
```

## Separated By Genotype

### Day 120

```{r}
D120_genotype_seurat <- subset(sub_seurat, timepoint == "D120")
D120_genotype_seurat <- SplitObject(D120_genotype_seurat, "genotype")

for (i in seq_along(D120_genotype_seurat)) {
  DefaultAssay(D120_genotype_seurat[[i]]) <- "RNA"
  D120_genotype_seurat[[i]][["SCT"]] <- NULL
  save_name <- paste(
    sub_name, unique(D120_genotype_seurat[[i]]$timepoint),
    unique(D120_genotype_seurat[[i]]$genotype), sep = "_"
  )
  D120_genotype_seurat[[i]] <- cluster_pipeline(
    D120_genotype_seurat[[i]], cache_dir, save_name, protocol,
    vars_to_regress, parallel_override, FALSE
  )
  print(D120_genotype_seurat[[i]])

  for (j in 1:(length(metadata_to_plot))) {
    print(names(metadata_to_plot)[j])
    if ("no_legend" %in% metadata_to_plot[[j]] && "no_label" %in% metadata_to_plot[[j]]) {
      print(
        red_dim_plot(D120_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j]) +
          NoLegend()
      )
    } else if ("no_legend" %in% metadata_to_plot[[j]]) {
      print(
        red_dim_plot(
          D120_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j], "cat"
        ) +
          NoLegend()
      )
    } else if ("no_label" %in% metadata_to_plot[[j]]) {
      print(red_dim_plot(D120_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j]))
    } else {
      print(
        red_dim_plot(D120_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j], "cat")
      )
    }
  }
}
```

### Day 72

```{r}
D72_genotype_seurat <- subset(sub_seurat, timepoint == "D72")
D72_genotype_seurat <- SplitObject(D72_genotype_seurat, "genotype")

for (i in seq_along(D72_genotype_seurat)) {
  DefaultAssay(D72_genotype_seurat[[i]]) <- "RNA"
  D72_genotype_seurat[[i]][["SCT"]] <- NULL
  save_name <- paste(
    sub_name, unique(D72_genotype_seurat[[i]]$timepoint),
    unique(D72_genotype_seurat[[i]]$genotype), sep = "_"
  )
  D72_genotype_seurat[[i]] <- cluster_pipeline(
    D72_genotype_seurat[[i]], cache_dir, save_name, protocol,
    vars_to_regress, parallel_override, FALSE
  )
  print(D72_genotype_seurat[[i]])

  for (j in 1:(length(metadata_to_plot))) {
    print(names(metadata_to_plot)[j])
    if ("no_legend" %in% metadata_to_plot[[j]] && "no_label" %in% metadata_to_plot[[j]]) {
      print(
        red_dim_plot(D72_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j]) +
          NoLegend()
      )
    } else if ("no_legend" %in% metadata_to_plot[[j]]) {
      print(
        red_dim_plot(
          D72_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j], "cat"
        ) +
          NoLegend()
      )
    } else if ("no_label" %in% metadata_to_plot[[j]]) {
      print(red_dim_plot(D72_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j]))
    } else {
      print(
        red_dim_plot(D72_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j], "cat")
      )
    }
  }
}
```

## Annotations

```{r}
sub_seurat <- RenameIdents(
  sub_seurat,
  "2" = "IPC 01", "0" = "IPC 02", "9" = "IPC 03",
  "1" = "IPC 04", "12" = "IPC 05", "17" = "IPC 06",
  "3" = "IPC 07", "4" = "IPC 08", "7" = "IPC 09",
  "13" = "IPC 10", "8" = "IPC 11", "10" = "IPC 12",
  "14" = "IPC 13", "16" = "IPC 14", "15" = "IPC 15",
  "11" = "IPC 16", "5" = "IPC 17", "6" = "IPC 18"
)
sub_seurat$clustermap_names <- factor(as.character(sub_seurat@active.ident))
sub_seurat$clustermap_names_short <- factor(
  sub("\\s+[^ ]+$", "", as.character(sub_seurat$clustermap_names))
)
red_dim_plot(sub_seurat, "umap1", "umap2", "clustermap_names", "cat") + ggtitle("D120 + D72")
red_dim_plot(sub_seurat, "umap1", "umap2", "clustermap_names_short", "cat") + ggtitle("D120 + D72")

timepoint_seurat$D120$clustermap_names <- factor(
  sub_seurat$clustermap_names[1:ncol(timepoint_seurat$D120)]
)
timepoint_seurat$D120$clustermap_names_short <- factor(
  sub_seurat$clustermap_names_short[1:ncol(timepoint_seurat$D120)]
)
red_dim_plot(timepoint_seurat$D120, "umap1", "umap2", "clustermap_names", "cat") + ggtitle("D120")
red_dim_plot(timepoint_seurat$D120, "umap1", "umap2", "clustermap_names_short", "cat") +
  ggtitle("D120")

timepoint_seurat$D72$clustermap_names <- factor(
  sub_seurat$clustermap_names[(ncol(timepoint_seurat$D120) + 1):ncol(sub_seurat)]
)
timepoint_seurat$D72$clustermap_names_short <- factor(
  sub_seurat$clustermap_names_short[(ncol(timepoint_seurat$D120) + 1):ncol(sub_seurat)]
)
red_dim_plot(timepoint_seurat$D72, "umap1", "umap2", "clustermap_names", "cat") + ggtitle("D72")
red_dim_plot(timepoint_seurat$D72, "umap1", "umap2", "clustermap_names_short", "cat") +
  ggtitle("D72")

file_list <- list(D120 = timepoint_seurat$D120, D72 = timepoint_seurat$D72, comb = sub_seurat)
rds <- file.path(cache_dir, paste(sub_name, "seurat_list.rds", sep = "_"))
if (!file.exists(rds)) {
  saveRDS(file_list, rds, compress = FALSE)
}

final_annotations <- c(final_annotations, as.character(file_list$comb$clustermap_names_short))
final_annotations_long <- c(final_annotations_long, as.character(file_list$comb$clustermap_names))
cell_order <- c(cell_order, colnames(sub_seurat))
```

# Inhibitory Neuron

## Standard Pipeline

```{r}
sub_seurat <- seurat[ , which(seurat$cluster_names == "Inhibitory Neuron")]

DefaultAssay(sub_seurat) <- "RNA"
sub_seurat[["SCT"]] <- NULL
sub_name <- "In"
sub_seurat <- cluster_pipeline(
  sub_seurat, cache_dir, sub_name, protocol, vars_to_regress, parallel_override, FALSE
)
sub_seurat

for (i in 1:length(metadata_to_plot)) {
  print(names(metadata_to_plot)[i])
  if ("no_legend" %in% metadata_to_plot[[i]] && "no_label" %in% metadata_to_plot[[i]]) {
    print(red_dim_plot(sub_seurat, "umap1", "umap2", names(metadata_to_plot)[i]) + NoLegend())
  } else if ("no_legend" %in% metadata_to_plot[[i]]) {
    print(
      red_dim_plot(sub_seurat, "umap1", "umap2", names(metadata_to_plot)[i], "cat") + NoLegend()
    )
  } else if ("no_label" %in% metadata_to_plot[[i]]) {
    print(red_dim_plot(sub_seurat, "umap1", "umap2", names(metadata_to_plot)[i]))
  } else {
    print(red_dim_plot(sub_seurat, "umap1", "umap2", names(metadata_to_plot)[i], "cat"))
  }
}

red_dim_plot(sub_seurat, "umap1", "umap2", "subsets_mito_percent", "cont") +
  ggtitle("Mito Percentage")
red_dim_plot(sub_seurat, "umap1", "umap2", "subsets_ribo_percent", "cont") +
  ggtitle("Ribo Percentage")
```

## Separated By Timepoint

```{r}
timepoint_seurat <- SplitObject(sub_seurat, "timepoint")

for (i in seq_along(timepoint_seurat)) {
  DefaultAssay(timepoint_seurat[[i]]) <- "RNA"
  timepoint_seurat[[i]][["SCT"]] <- NULL
  save_name <- paste(sub_name, unique(timepoint_seurat[[i]]$timepoint), sep = "_")
  timepoint_seurat[[i]] <- cluster_pipeline(
    timepoint_seurat[[i]], cache_dir, save_name, protocol, vars_to_regress, parallel_override, FALSE
  )
  print(timepoint_seurat[[i]])

  for (j in 1:(length(metadata_to_plot))) {
    print(names(metadata_to_plot)[j])
    if ("no_legend" %in% metadata_to_plot[[j]] && "no_label" %in% metadata_to_plot[[j]]) {
      print(
        red_dim_plot(timepoint_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j]) +
          NoLegend()
      )
    } else if ("no_legend" %in% metadata_to_plot[[j]]) {
      print(
        red_dim_plot(timepoint_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j], "cat") +
          NoLegend()
      )
    } else if ("no_label" %in% metadata_to_plot[[j]]) {
      print(red_dim_plot(timepoint_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j]))
    } else {
      print(
        red_dim_plot(timepoint_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j], "cat")
      )
    }
  }

  for (j in seq_along(markers)) {
    if (any(markers[[j]] %in% rownames(timepoint_seurat[[i]]))) {
      print(
        FeaturePlot(
          timepoint_seurat[[i]], markers[[j]],
          order = TRUE, cols = c("lightgrey", "red")
        ) +
          plot_annotation(
            names(markers)[j],
            theme = theme(plot.title = element_text(hjust = 0.5, size = 18, face = "bold"))
          )
      )
    }
  }
}
```

## Separated By Genotype

### Day 120

```{r}
D120_genotype_seurat <- subset(sub_seurat, timepoint == "D120")
D120_genotype_seurat <- SplitObject(D120_genotype_seurat, "genotype")

for (i in seq_along(D120_genotype_seurat)) {
  DefaultAssay(D120_genotype_seurat[[i]]) <- "RNA"
  D120_genotype_seurat[[i]][["SCT"]] <- NULL
  save_name <- paste(
    sub_name, unique(D120_genotype_seurat[[i]]$timepoint),
    unique(D120_genotype_seurat[[i]]$genotype), sep = "_"
  )
  D120_genotype_seurat[[i]] <- cluster_pipeline(
    D120_genotype_seurat[[i]], cache_dir, save_name, protocol,
    vars_to_regress, parallel_override, FALSE
  )
  print(D120_genotype_seurat[[i]])

  for (j in 1:(length(metadata_to_plot))) {
    print(names(metadata_to_plot)[j])
    if ("no_legend" %in% metadata_to_plot[[j]] && "no_label" %in% metadata_to_plot[[j]]) {
      print(
        red_dim_plot(D120_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j]) +
          NoLegend()
      )
    } else if ("no_legend" %in% metadata_to_plot[[j]]) {
      print(
        red_dim_plot(
          D120_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j], "cat"
        ) +
          NoLegend()
      )
    } else if ("no_label" %in% metadata_to_plot[[j]]) {
      print(red_dim_plot(D120_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j]))
    } else {
      print(
        red_dim_plot(D120_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j], "cat")
      )
    }
  }
}
```

### Day 72

```{r}
D72_genotype_seurat <- subset(sub_seurat, timepoint == "D72")
D72_genotype_seurat <- SplitObject(D72_genotype_seurat, "genotype")

for (i in seq_along(D72_genotype_seurat)) {
  DefaultAssay(D72_genotype_seurat[[i]]) <- "RNA"
  D72_genotype_seurat[[i]][["SCT"]] <- NULL
  save_name <- paste(
    sub_name, unique(D72_genotype_seurat[[i]]$timepoint),
    unique(D72_genotype_seurat[[i]]$genotype), sep = "_"
  )
  D72_genotype_seurat[[i]] <- cluster_pipeline(
    D72_genotype_seurat[[i]], cache_dir, save_name, protocol,
    vars_to_regress, parallel_override, FALSE
  )
  print(D72_genotype_seurat[[i]])

  for (j in 1:(length(metadata_to_plot))) {
    print(names(metadata_to_plot)[j])
    if ("no_legend" %in% metadata_to_plot[[j]] && "no_label" %in% metadata_to_plot[[j]]) {
      print(
        red_dim_plot(D72_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j]) +
          NoLegend()
      )
    } else if ("no_legend" %in% metadata_to_plot[[j]]) {
      print(
        red_dim_plot(
          D72_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j], "cat"
        ) +
          NoLegend()
      )
    } else if ("no_label" %in% metadata_to_plot[[j]]) {
      print(red_dim_plot(D72_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j]))
    } else {
      print(
        red_dim_plot(D72_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j], "cat")
      )
    }
  }
}
```

## Annotations

```{r}
sub_seurat <- RenameIdents(
  sub_seurat,
  "14" = "In Unclassified 01", "8" = "In Unclassified 02", "7" = "In Unclassified 03",
  "13" = "In Unclassified 04", "18" = "In Unclassified 05", "22" = "In Unclassified 06",
  "24" = "In Unclassified 07", "1" = "In 01", "4" = "In 02",
  "2" = "In 03", "19" = "In 04", "10" = "In 05",
  "12" = "In 06", "11" = "In 07", "3" = "In 08",
  "20" = "In 09", "16" = "In 10", "15" = "In 11",
  "0" = "In 12", "23" = "In 13", "9" = "In 14",
  "17" = "In 15", "5" = "In 16", "21" = "In 17",
  "6" = "In 18"
)
sub_seurat$clustermap_names <- factor(as.character(sub_seurat@active.ident))
sub_seurat$clustermap_names_short <- factor(
  sub("\\s+[^ ]+$", "", as.character(sub_seurat$clustermap_names))
)
red_dim_plot(sub_seurat, "umap1", "umap2", "clustermap_names", "cat") + ggtitle("D120 + D72")
red_dim_plot(sub_seurat, "umap1", "umap2", "clustermap_names_short", "cat") + ggtitle("D120 + D72")

timepoint_seurat$D120$clustermap_names <- factor(
  sub_seurat$clustermap_names[1:ncol(timepoint_seurat$D120)]
)
timepoint_seurat$D120$clustermap_names_short <- factor(
  sub_seurat$clustermap_names_short[1:ncol(timepoint_seurat$D120)]
)
red_dim_plot(timepoint_seurat$D120, "umap1", "umap2", "clustermap_names", "cat") + ggtitle("D120")
red_dim_plot(timepoint_seurat$D120, "umap1", "umap2", "clustermap_names_short", "cat") +
  ggtitle("D120")

timepoint_seurat$D72$clustermap_names <- factor(
  sub_seurat$clustermap_names[(ncol(timepoint_seurat$D120) + 1):ncol(sub_seurat)]
)
timepoint_seurat$D72$clustermap_names_short <- factor(
  sub_seurat$clustermap_names_short[(ncol(timepoint_seurat$D120) + 1):ncol(sub_seurat)]
)
red_dim_plot(timepoint_seurat$D72, "umap1", "umap2", "clustermap_names", "cat") + ggtitle("D72")
red_dim_plot(timepoint_seurat$D72, "umap1", "umap2", "clustermap_names_short", "cat") +
  ggtitle("D72")

file_list <- list(D120 = timepoint_seurat$D120, D72 = timepoint_seurat$D72, comb = sub_seurat)
rds <- file.path(cache_dir, paste(sub_name, "seurat_list.rds", sep = "_"))
if (!file.exists(rds)) {
  saveRDS(file_list, rds, compress = FALSE)
}

final_annotations <- c(final_annotations, as.character(file_list$comb$clustermap_names_short))
final_annotations_long <- c(final_annotations_long, as.character(file_list$comb$clustermap_names))
cell_order <- c(cell_order, colnames(sub_seurat))
```

# Mural

## Standard Pipeline

```{r}
sub_seurat <- seurat[ , which(seurat$cluster_names == "Mural")]

DefaultAssay(sub_seurat) <- "RNA"
sub_seurat[["SCT"]] <- NULL
sub_name <- "Mu"
sub_seurat <- cluster_pipeline(
  sub_seurat, cache_dir, sub_name, protocol, vars_to_regress, parallel_override, FALSE
)
sub_seurat

for (i in 1:length(metadata_to_plot)) {
  print(names(metadata_to_plot)[i])
  if ("no_legend" %in% metadata_to_plot[[i]] && "no_label" %in% metadata_to_plot[[i]]) {
    print(red_dim_plot(sub_seurat, "umap1", "umap2", names(metadata_to_plot)[i]) + NoLegend())
  } else if ("no_legend" %in% metadata_to_plot[[i]]) {
    print(
      red_dim_plot(sub_seurat, "umap1", "umap2", names(metadata_to_plot)[i], "cat") + NoLegend()
    )
  } else if ("no_label" %in% metadata_to_plot[[i]]) {
    print(red_dim_plot(sub_seurat, "umap1", "umap2", names(metadata_to_plot)[i]))
  } else {
    print(red_dim_plot(sub_seurat, "umap1", "umap2", names(metadata_to_plot)[i], "cat"))
  }
}

red_dim_plot(sub_seurat, "umap1", "umap2", "subsets_mito_percent", "cont") +
  ggtitle("Mito Percentage")
red_dim_plot(sub_seurat, "umap1", "umap2", "subsets_ribo_percent", "cont") +
  ggtitle("Ribo Percentage")
```

## Separated By Timepoint

```{r}
timepoint_seurat <- SplitObject(sub_seurat, "timepoint")

for (i in seq_along(timepoint_seurat)) {
  DefaultAssay(timepoint_seurat[[i]]) <- "RNA"
  timepoint_seurat[[i]][["SCT"]] <- NULL
  save_name <- paste(sub_name, unique(timepoint_seurat[[i]]$timepoint), sep = "_")
  timepoint_seurat[[i]] <- cluster_pipeline(
    timepoint_seurat[[i]], cache_dir, save_name, protocol, vars_to_regress, parallel_override, FALSE
  )
  print(timepoint_seurat[[i]])

  for (j in 1:(length(metadata_to_plot))) {
    print(names(metadata_to_plot)[j])
    if ("no_legend" %in% metadata_to_plot[[j]] && "no_label" %in% metadata_to_plot[[j]]) {
      print(
        red_dim_plot(timepoint_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j]) +
          NoLegend()
      )
    } else if ("no_legend" %in% metadata_to_plot[[j]]) {
      print(
        red_dim_plot(timepoint_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j], "cat") +
          NoLegend()
      )
    } else if ("no_label" %in% metadata_to_plot[[j]]) {
      print(red_dim_plot(timepoint_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j]))
    } else {
      print(
        red_dim_plot(timepoint_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j], "cat")
      )
    }
  }

  for (j in seq_along(markers)) {
    if (any(markers[[j]] %in% rownames(timepoint_seurat[[i]]))) {
      print(
        FeaturePlot(
          timepoint_seurat[[i]], markers[[j]],
          order = TRUE, cols = c("lightgrey", "red")
        ) +
          plot_annotation(
            names(markers)[j],
            theme = theme(plot.title = element_text(hjust = 0.5, size = 18, face = "bold"))
          )
      )
    }
  }
}
```

## Separated By Genotype

### Day 120

```{r}
D120_genotype_seurat <- subset(sub_seurat, timepoint == "D120")
D120_genotype_seurat <- SplitObject(D120_genotype_seurat, "genotype")

for (i in seq_along(D120_genotype_seurat)) {
  DefaultAssay(D120_genotype_seurat[[i]]) <- "RNA"
  D120_genotype_seurat[[i]][["SCT"]] <- NULL
  save_name <- paste(
    sub_name, unique(D120_genotype_seurat[[i]]$timepoint),
    unique(D120_genotype_seurat[[i]]$genotype), sep = "_"
  )
  D120_genotype_seurat[[i]] <- cluster_pipeline(
    D120_genotype_seurat[[i]], cache_dir, save_name, protocol,
    vars_to_regress, parallel_override, FALSE
  )
  print(D120_genotype_seurat[[i]])

  for (j in 1:(length(metadata_to_plot))) {
    print(names(metadata_to_plot)[j])
    if ("no_legend" %in% metadata_to_plot[[j]] && "no_label" %in% metadata_to_plot[[j]]) {
      print(
        red_dim_plot(D120_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j]) +
          NoLegend()
      )
    } else if ("no_legend" %in% metadata_to_plot[[j]]) {
      print(
        red_dim_plot(
          D120_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j], "cat"
        ) +
          NoLegend()
      )
    } else if ("no_label" %in% metadata_to_plot[[j]]) {
      print(red_dim_plot(D120_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j]))
    } else {
      print(
        red_dim_plot(D120_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j], "cat")
      )
    }
  }
}
```

### Day 72

```{r}
D72_genotype_seurat <- subset(sub_seurat, timepoint == "D72")
D72_genotype_seurat <- SplitObject(D72_genotype_seurat, "genotype")

for (i in seq_along(D72_genotype_seurat)) {
  DefaultAssay(D72_genotype_seurat[[i]]) <- "RNA"
  D72_genotype_seurat[[i]][["SCT"]] <- NULL
  save_name <- paste(
    sub_name, unique(D72_genotype_seurat[[i]]$timepoint),
    unique(D72_genotype_seurat[[i]]$genotype), sep = "_"
  )
  D72_genotype_seurat[[i]] <- cluster_pipeline(
    D72_genotype_seurat[[i]], cache_dir, save_name, protocol,
    vars_to_regress, parallel_override, FALSE
  )
  print(D72_genotype_seurat[[i]])

  for (j in 1:(length(metadata_to_plot))) {
    print(names(metadata_to_plot)[j])
    if ("no_legend" %in% metadata_to_plot[[j]] && "no_label" %in% metadata_to_plot[[j]]) {
      print(
        red_dim_plot(D72_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j]) +
          NoLegend()
      )
    } else if ("no_legend" %in% metadata_to_plot[[j]]) {
      print(
        red_dim_plot(
          D72_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j], "cat"
        ) +
          NoLegend()
      )
    } else if ("no_label" %in% metadata_to_plot[[j]]) {
      print(red_dim_plot(D72_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j]))
    } else {
      print(
        red_dim_plot(D72_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j], "cat")
      )
    }
  }
}
```

## Annotations

```{r}
sub_seurat <- RenameIdents(
  sub_seurat,
  "0" = "Mu 01", "3" = "Mu 02", "4" = "Mu 03",
  "2" = "Mu 04", "1" = "Mu 05"
)
sub_seurat$clustermap_names <- factor(as.character(sub_seurat@active.ident))
sub_seurat$clustermap_names_short <- factor(
  sub("\\s+[^ ]+$", "", as.character(sub_seurat$clustermap_names))
)
red_dim_plot(sub_seurat, "umap1", "umap2", "clustermap_names", "cat") + ggtitle("D120 + D72")
red_dim_plot(sub_seurat, "umap1", "umap2", "clustermap_names_short", "cat") + ggtitle("D120 + D72")

timepoint_seurat$D120$clustermap_names <- factor(
  sub_seurat$clustermap_names[1:ncol(timepoint_seurat$D120)]
)
timepoint_seurat$D120$clustermap_names_short <- factor(
  sub_seurat$clustermap_names_short[1:ncol(timepoint_seurat$D120)]
)
red_dim_plot(timepoint_seurat$D120, "umap1", "umap2", "clustermap_names", "cat") + ggtitle("D120")
red_dim_plot(timepoint_seurat$D120, "umap1", "umap2", "clustermap_names_short", "cat") +
  ggtitle("D120")

timepoint_seurat$D72$clustermap_names <- factor(
  sub_seurat$clustermap_names[(ncol(timepoint_seurat$D120) + 1):ncol(sub_seurat)]
)
timepoint_seurat$D72$clustermap_names_short <- factor(
  sub_seurat$clustermap_names_short[(ncol(timepoint_seurat$D120) + 1):ncol(sub_seurat)]
)
red_dim_plot(timepoint_seurat$D72, "umap1", "umap2", "clustermap_names", "cat") + ggtitle("D72")
red_dim_plot(timepoint_seurat$D72, "umap1", "umap2", "clustermap_names_short", "cat") +
  ggtitle("D72")

file_list <- list(D120 = timepoint_seurat$D120, D72 = timepoint_seurat$D72, comb = sub_seurat)
rds <- file.path(cache_dir, paste(sub_name, "seurat_list.rds", sep = "_"))
if (!file.exists(rds)) {
  saveRDS(file_list, rds, compress = FALSE)
}

final_annotations <- c(final_annotations, as.character(file_list$comb$clustermap_names_short))
final_annotations_long <- c(final_annotations_long, as.character(file_list$comb$clustermap_names))
cell_order <- c(cell_order, colnames(sub_seurat))
```

# Unclassified

## Standard Pipeline

```{r}
sub_seurat <- seurat[ , which(seurat$cluster_names == "Unclassified")]

DefaultAssay(sub_seurat) <- "RNA"
sub_seurat[["SCT"]] <- NULL
sub_name <- "Unc"
sub_seurat <- cluster_pipeline(
  sub_seurat, cache_dir, sub_name, protocol, vars_to_regress, parallel_override, FALSE
)
sub_seurat

for (i in 1:length(metadata_to_plot)) {
  print(names(metadata_to_plot)[i])
  if ("no_legend" %in% metadata_to_plot[[i]] && "no_label" %in% metadata_to_plot[[i]]) {
    print(red_dim_plot(sub_seurat, "umap1", "umap2", names(metadata_to_plot)[i]) + NoLegend())
  } else if ("no_legend" %in% metadata_to_plot[[i]]) {
    print(
      red_dim_plot(sub_seurat, "umap1", "umap2", names(metadata_to_plot)[i], "cat") + NoLegend()
    )
  } else if ("no_label" %in% metadata_to_plot[[i]]) {
    print(red_dim_plot(sub_seurat, "umap1", "umap2", names(metadata_to_plot)[i]))
  } else {
    print(red_dim_plot(sub_seurat, "umap1", "umap2", names(metadata_to_plot)[i], "cat"))
  }
}

red_dim_plot(sub_seurat, "umap1", "umap2", "subsets_mito_percent", "cont") +
  ggtitle("Mito Percentage")
red_dim_plot(sub_seurat, "umap1", "umap2", "subsets_ribo_percent", "cont") +
  ggtitle("Ribo Percentage")
```

## Separated By Timepoint

```{r}
timepoint_seurat <- SplitObject(sub_seurat, "timepoint")

for (i in seq_along(timepoint_seurat)) {
  DefaultAssay(timepoint_seurat[[i]]) <- "RNA"
  timepoint_seurat[[i]][["SCT"]] <- NULL
  save_name <- paste(sub_name, unique(timepoint_seurat[[i]]$timepoint), sep = "_")
  timepoint_seurat[[i]] <- cluster_pipeline(
    timepoint_seurat[[i]], cache_dir, save_name, protocol, vars_to_regress, parallel_override, FALSE
  )
  print(timepoint_seurat[[i]])

  for (j in 1:(length(metadata_to_plot))) {
    print(names(metadata_to_plot)[j])
    if ("no_legend" %in% metadata_to_plot[[j]] && "no_label" %in% metadata_to_plot[[j]]) {
      print(
        red_dim_plot(timepoint_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j]) +
          NoLegend()
      )
    } else if ("no_legend" %in% metadata_to_plot[[j]]) {
      print(
        red_dim_plot(timepoint_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j], "cat") +
          NoLegend()
      )
    } else if ("no_label" %in% metadata_to_plot[[j]]) {
      print(red_dim_plot(timepoint_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j]))
    } else {
      print(
        red_dim_plot(timepoint_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j], "cat")
      )
    }
  }

  for (j in seq_along(markers)) {
    if (any(markers[[j]] %in% rownames(timepoint_seurat[[i]]))) {
      print(
        FeaturePlot(
          timepoint_seurat[[i]], markers[[j]],
          order = TRUE, cols = c("lightgrey", "red")
        ) +
          plot_annotation(
            names(markers)[j],
            theme = theme(plot.title = element_text(hjust = 0.5, size = 18, face = "bold"))
          )
      )
    }
  }
}
```

## Separated By Genotype

### Day 120

```{r}
D120_genotype_seurat <- subset(sub_seurat, timepoint == "D120")
D120_genotype_seurat <- SplitObject(D120_genotype_seurat, "genotype")

for (i in seq_along(D120_genotype_seurat)) {
  DefaultAssay(D120_genotype_seurat[[i]]) <- "RNA"
  D120_genotype_seurat[[i]][["SCT"]] <- NULL
  save_name <- paste(
    sub_name, unique(D120_genotype_seurat[[i]]$timepoint),
    unique(D120_genotype_seurat[[i]]$genotype), sep = "_"
  )
  D120_genotype_seurat[[i]] <- cluster_pipeline(
    D120_genotype_seurat[[i]], cache_dir, save_name, protocol,
    vars_to_regress, parallel_override, FALSE
  )
  print(D120_genotype_seurat[[i]])

  for (j in 1:(length(metadata_to_plot))) {
    print(names(metadata_to_plot)[j])
    if ("no_legend" %in% metadata_to_plot[[j]] && "no_label" %in% metadata_to_plot[[j]]) {
      print(
        red_dim_plot(D120_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j]) +
          NoLegend()
      )
    } else if ("no_legend" %in% metadata_to_plot[[j]]) {
      print(
        red_dim_plot(
          D120_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j], "cat"
        ) +
          NoLegend()
      )
    } else if ("no_label" %in% metadata_to_plot[[j]]) {
      print(red_dim_plot(D120_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j]))
    } else {
      print(
        red_dim_plot(D120_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j], "cat")
      )
    }
  }
}
```

### Day 72

```{r}
D72_genotype_seurat <- subset(sub_seurat, timepoint == "D72")
D72_genotype_seurat <- SplitObject(D72_genotype_seurat, "genotype")

for (i in seq_along(D72_genotype_seurat)) {
  DefaultAssay(D72_genotype_seurat[[i]]) <- "RNA"
  D72_genotype_seurat[[i]][["SCT"]] <- NULL
  save_name <- paste(
    sub_name, unique(D72_genotype_seurat[[i]]$timepoint),
    unique(D72_genotype_seurat[[i]]$genotype), sep = "_"
  )
  D72_genotype_seurat[[i]] <- cluster_pipeline(
    D72_genotype_seurat[[i]], cache_dir, save_name, protocol,
    vars_to_regress, parallel_override, FALSE
  )
  print(D72_genotype_seurat[[i]])

  for (j in 1:(length(metadata_to_plot))) {
    print(names(metadata_to_plot)[j])
    if ("no_legend" %in% metadata_to_plot[[j]] && "no_label" %in% metadata_to_plot[[j]]) {
      print(
        red_dim_plot(D72_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j]) +
          NoLegend()
      )
    } else if ("no_legend" %in% metadata_to_plot[[j]]) {
      print(
        red_dim_plot(
          D72_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j], "cat"
        ) +
          NoLegend()
      )
    } else if ("no_label" %in% metadata_to_plot[[j]]) {
      print(red_dim_plot(D72_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j]))
    } else {
      print(
        red_dim_plot(D72_genotype_seurat[[i]], "umap1", "umap2", names(metadata_to_plot)[j], "cat")
      )
    }
  }
}
```

## Annotations

```{r}
sub_seurat <- RenameIdents(
  sub_seurat,
  "0" = "Unc 01", "6" = "Unc 02", "1" = "Unc 03",
  "3" = "Unc 04", "4" = "Unc 05", "10" = "Unc 06",
  "9" = "Unc 07", "2" = "Unc 08", "8" = "Unc 09",
  "7" = "Unc 10", "5" = "Unc 11"
)
sub_seurat$clustermap_names <- factor(as.character(sub_seurat@active.ident))
sub_seurat$clustermap_names_short <- factor(
  sub("\\s+[^ ]+$", "", as.character(sub_seurat$clustermap_names))
)
red_dim_plot(sub_seurat, "umap1", "umap2", "clustermap_names", "cat") + ggtitle("D120 + D72")
red_dim_plot(sub_seurat, "umap1", "umap2", "clustermap_names_short", "cat") + ggtitle("D120 + D72")

timepoint_seurat$D120$clustermap_names <- factor(
  sub_seurat$clustermap_names[1:ncol(timepoint_seurat$D120)]
)
timepoint_seurat$D120$clustermap_names_short <- factor(
  sub_seurat$clustermap_names_short[1:ncol(timepoint_seurat$D120)]
)
red_dim_plot(timepoint_seurat$D120, "umap1", "umap2", "clustermap_names", "cat") + ggtitle("D120")
red_dim_plot(timepoint_seurat$D120, "umap1", "umap2", "clustermap_names_short", "cat") +
  ggtitle("D120")

timepoint_seurat$D72$clustermap_names <- factor(
  sub_seurat$clustermap_names[(ncol(timepoint_seurat$D120) + 1):ncol(sub_seurat)]
)
timepoint_seurat$D72$clustermap_names_short <- factor(
  sub_seurat$clustermap_names_short[(ncol(timepoint_seurat$D120) + 1):ncol(sub_seurat)]
)
red_dim_plot(timepoint_seurat$D72, "umap1", "umap2", "clustermap_names", "cat") + ggtitle("D72")
red_dim_plot(timepoint_seurat$D72, "umap1", "umap2", "clustermap_names_short", "cat") +
  ggtitle("D72")

file_list <- list(D120 = timepoint_seurat$D120, D72 = timepoint_seurat$D72, comb = sub_seurat)
rds <- file.path(cache_dir, paste(sub_name, "seurat_list.rds", sep = "_"))
if (!file.exists(rds)) {
  saveRDS(file_list, rds, compress = FALSE)
}

final_annotations <- c(final_annotations, as.character(file_list$comb$clustermap_names_short))
final_annotations_long <- c(final_annotations_long, as.character(file_list$comb$clustermap_names))
cell_order <- c(cell_order, colnames(sub_seurat))
```

# Final Annotations

```{r}
final_annotations <- final_annotations[match(colnames(seurat), cell_order)]
final_annotations_long <- final_annotations_long[match(colnames(seurat), cell_order)]

seurat$subclusters <- factor(
  final_annotations,
  c(
    "Ex Deep Layer", "Ex Upper Layer", "Ex Newborn", "Ex Unclassified", "In",
    "In Unclassified", "RG Astrocyte", "RG OPC", "RG Early", "RG Late",
    "RG Unclassified", "IPC", "Mu", "Unc"
  )
)
seurat$subclusters_long <- factor(final_annotations_long)
n <- length(unique(seurat$subclusters))
color <- cols25(n + 3)[c(1:(n - 2), n + 3, n)]
red_dim_plot(seurat, "rca_umap1", "rca_umap2", "subclusters") +
  guides(color = guide_legend(override.aes = list(size = 10))) +
  scale_fill_manual(values = color)
red_dim_plot(seurat, "rca_umap1", "rca_umap2", "subclusters_long") + NoLegend()

rds <- file.path(cache_dir, "seurat.rds")
if (!file.exists(rds)) {
  saveRDS(seurat, rds, compress = FALSE)
}
```

# References

This is the concluding section of the document. Here we output the `sessionInfo` and create a bibliography for works cited.

```{r}
sessionInfo()
```
